requirements.txt
flask
flask-cors
chromadb
openai
python-dotenv
